requirements.txt
streamlit
pandas
python-dotenv
openai
google-generativeai
